+++ python/cuml/linear_model/linear_regression.pyx
@@ -24,11 +24,11 @@ import cudf
 import numpy as np
 
 from numba import cuda
-
+from collections import defaultdict
 from libcpp cimport bool
 from libc.stdint cimport uintptr_t
 from libc.stdlib cimport calloc, malloc, free
-
+from sklearn.utils.fixes import signature
 
 cdef extern from "glm/glm_c.h" namespace "ML::GLM":
 
@@ -71,8 +71,9 @@ class LinearRegression:
     LinearRegression is a simple machine learning model where the response y is modelled by a 
     linear combination of the predictors in X.
 
-    cuML's LinearRegression expects a cuDF DataFrame, and provides 2 algorithms SVD and Eig to 
-    fit a linear model. SVD is more stable, but Eig (default) is much more faster.
+    cuML's LinearRegression expects either a cuDF DataFrame or a NumPy matrix and provides 2 
+    algorithms SVD and Eig to fit a linear model. SVD is more stable, but Eig (default) 
+    is much more faster.
 
     Examples
     --------
@@ -147,7 +148,6 @@ class LinearRegression:
     For additional docs, see `scikitlearn's OLS <https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html>`_.
 
     """
-    #For an additional example see `the OLS notebook <https://github.com/rapidsai/cuml/blob/master/python/notebooks/glm_demo.ipynb>`_. 
 
     def __init__(self, algorithm='eig', fit_intercept=True, normalize=False):
 
@@ -247,6 +247,7 @@ class LinearRegression:
 
         cdef float c_intercept1
         cdef double c_intercept2
+
         if self.gdf_datatype.type == np.float32:
 
             olsFit(<float*>X_ptr,
@@ -272,8 +273,8 @@ class LinearRegression:
                        <int>self.algo)
 
             self.intercept_ = c_intercept2
-
         return self
+
 
     def predict(self, X):
         """
@@ -332,3 +333,32 @@ class LinearRegression:
         del(X_m)
 
         return preds
+
+      
+    def get_params(self, deep=True):
+        params = dict()
+        variables = ['algorithm','fit_intercept','normalize']
+        for key in variables:
+            var_value = getattr(self,key,None)
+            params[key] = var_value   
+        return params
+
+      
+    def set_params(self, **params):
+        if not params:
+            return self
+        variables = ['algorithm','fit_intercept','normalize']
+
+        current_params = {"algorithm" : self.algo,"fit_intercept" : self.fit_intercept,"normalize" : self.normalize}
+        for key, value in params.items():
+            if key not in current_params:
+                raise ValueError('Invalid parameter %s for estimator')
+            else:
+                setattr(self, key, value)
+                current_params[key] = value
+        if params["algorithm"]=='eig':
+            self.algo=1
+        else:
+            self.algo = 0
+        return self
+