+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 # cuML 0.6.0 (Date TBD)
 
 ## New Features
-
+- PR #235: NearestNeighbors MG Support
 
 
 ## Improvements
@@ -40,6 +40,7 @@
 - PR #232: Added reduce_rows_by_key
 - PR #246: Support for 2 vectors in the matrix vector operator
 - PR #244: Fix for single GPU OLS and Ridge to support one column training data
+- PR #271: Added get_params and set_params functions for linear and ridge regression
 - PR #253: Fix for issue #250-reduce_rows_by_key failed memcheck for small nkeys
 - PR #269: LinearRegression, Ridge Python docs update and cleaning